ADD: detect cycle rule + fix for cyclical state def.
