+++ src/components/ServerAddUser.tsx
@@ -1,9 +1,10 @@
+import type { IServerParams } from '~Types';
+
 import { IUser } from '@vex-chat/libvex';
 import React, { Fragment, useState } from 'react';
 import { useSelector } from 'react-redux';
 import { useHistory, useParams } from 'react-router';
 import { selectServers } from '../reducers/servers';
-import { IServerParams } from '../views/Server';
 import { IconUsername } from './IconUsername';
 import { emptyUser } from './MessagingBar';
 import { UserSearchBar } from './UserSearchBar';
