+++ src/rootReducer.ts
@@ -10,20 +10,16 @@ import channelReducer from './reducers/channels';
 import permissionsReducer from './reducers/permissions';
 import groupMessageReducer from './reducers/groupMessages';
 
-// eslint-disable-next-line @typescript-eslint/explicit-module-boundary-types
-export default function createRootReducer() {
-    const root = combineReducers({
-        user: userReducer,
-        familiars: familiarsReducer,
-        inputs: inputsReducer,
-        messages: messageReducer,
-        sessions: sessionReducer,
-        app: appReducer,
-        servers: serverReducer,
-        channels: channelReducer,
-        groupMessages: groupMessageReducer,
-        permissions: permissionsReducer,
-    });
 
-    return root;
-}
+export default combineReducers({
+    user: userReducer,
+    familiars: familiarsReducer,
+    inputs: inputsReducer,
+    messages: messageReducer,
+    sessions: sessionReducer,
+    app: appReducer,
+    servers: serverReducer,
+    channels: channelReducer,
+    groupMessages: groupMessageReducer,
+    permissions: permissionsReducer,
+});
