+++ src/utils/KeyGaurdian.ts
@@ -1,7 +1,20 @@
-import { Client } from '@vex-chat/libvex';
-
-export class KeyGaurdian {
+import { Client } from "@vex-chat/libvex";
+class KeyGaurdian {
+    private static instance: KeyGaurdian;
+    
     private SK: string | null = null;
+
+    // The constructor is empty because we need to make it private.
+    // eslint-disable-next-line @typescript-eslint/no-empty-function
+    private constructor() { }
+
+    public static getInstance(): KeyGaurdian {
+        if (!KeyGaurdian.instance) {
+            KeyGaurdian.instance = new KeyGaurdian();
+        }
+
+        return KeyGaurdian.instance;
+    }
 
     public hasKey(): boolean {
         return this.SK !== null;
@@ -9,7 +22,7 @@ export class KeyGaurdian {
 
     public getKey(): string {
         if (!this.SK) {
-            throw new Error('Must call load() before using this function.');
+            throw new Error("Must call load() before using this function.");
         }
         return this.SK;
     }
@@ -23,3 +36,7 @@ export class KeyGaurdian {
         this.SK = null;
     }
 }
+
+const guardian = KeyGaurdian.getInstance()
+
+export default guardian
