+++ src/views/Login.tsx
@@ -1,3 +1,5 @@
+import type { RootState } from '~Types';
+
 import { faLock } from '@fortawesome/free-solid-svg-icons';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import React, { useState, FunctionComponent, memo, useEffect } from 'react';
@@ -5,15 +7,16 @@ import { useSelector, useDispatch } from 'react-redux';
 import { useHistory } from 'react-router';
 import { routes } from '../constants/routes';
 import { addInputState } from '../reducers/inputs';
-import { gaurdian } from './Base';
+
 import { BackButton } from '../components/BackButton';
 import { keyFolder } from '../constants/folders';
 import { VerticalAligner } from '../components/VerticalAligner';
 import { useQuery } from '../hooks/useQuery';
 import { useDebounce } from '../hooks/useDebounce';
-import { RootState } from '../store';
+import gaurdian from '../utils/KeyGaurdian'
 
 const FORM_NAME = 'keyfile-login-pasword';
+
 
 export const Login: FunctionComponent = memo(() => {
     const history = useHistory();
